modified

+++ BasicMath/ReverseInteger.java
@@ -3,12 +3,12 @@ package BasicMath;
 import java.util.Scanner;
 
 public class ReverseInteger {
-    static int Valid(int x){
+    static long Valid(long x){
         int sign = x<0 ? -1 : 1;
         x=Math.abs(x);
-        int res=0;
+        long res=0;
         while(x>0){
-            int d=x%10;
+            int d=(int)x%10;
             res=(res*10)+d;
             x/=10;
 
@@ -18,7 +18,7 @@ public class ReverseInteger {
     }
 public static void main(String[] args) {
         Scanner sc=new Scanner(System.in);
-        int x=sc.nextInt();
+        long x=sc.nextInt();
         System.out.println(Valid(x));
         sc.close();
     }
